requirements.txt
aioredis==1.1.0
amqp==2.6.1
appdirs==1.4.4
apturl==0.5.2
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
awscli==1.18.69
beautifulsoup4==4.6.0
billiard==3.6.3.0
bleach==3.1.1
blinker==1.3
boto3==1.16.25
botocore==1.19.25
Brlapi==0.6.6
celery==4.3.0
certifi==2018.8.13
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
colorama==0.3.7
command-not-found==0.3
constantly==15.1.0
coverage==5.1
cryptography==2.1.4
cupshelpers==1.0
daphne==2.2.1
defer==1.0.6
defusedxml==0.5.0
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==2.1
django-allauth==0.36.0
django-bootstrap4==0.0.6
django-celery-results==1.1.2
django-cors-headers==2.4.0
django-extensions==2.1.6
django-guardian==2.1.0
django-mathfilters==0.4.0
django-organizations==1.1.1
django-otp==0.6.0
django-password-validators==0.1.7
django-post-office==3.0.4
django-rest-auth==0.9.5
django-rosetta==0.9.3
django-storages==1.9.1
djangorestframework==3.8.2
dnslib==0.9.14
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
feedparser==5.2.1
filelock==3.0.12
hiredis==0.2.0
html5lib==0.999999999
httplib2==0.9.1
hyperlink==18.0.0
idna==2.7
imgkit==1.0.1
importlib-metadata==2.0.0
importlib-resources==3.0.0
incremental==17.5.0
isort==5.6.3
jmespath==0.9.3
jsonfield==2.0.2
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
localstack==0.12.2
localstack-client==1.7
localstack-ext==0.12.0
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mock==4.0.1
msgpack==0.5.6
netifaces==0.10.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
pdfkit==0.6.1
pexpect==4.0.1
Pillow==7.2.0
polib==1.1.0
protobuf==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.5
pyaes==1.6.1
pyaml==17.12.1
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.5.3
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pypandoc==1.5
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.7.3
python-debian==0.1.32
python-memcached==1.59
python-pdf==0.36
python3-openid==3.1.0
pytz==2018.5
pyxdg==0.25
PyYAML==3.13
redis==3.2.1
reportlab==3.3.0
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sqlparse==0.2.4
stripe==2.24.1
system-service==0.3
systemd-python==234
toml==0.10.1
Twisted==18.7.0
txaio==18.7.1
typed-ast==1.4.1
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.26.2
usb-creator==0.3.3
vine==5.0.0
virtualenv==20.0.33
wadllib==1.3.2
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==3.3.0
zope.interface==5.1.0
